The progress bar can display the completion  rate of task.

+++ LiTsap/app/src/main/java/studio/honidot/litsap/data/Task.kt
@@ -6,8 +6,8 @@ data class Task(
     val id: Long,
     val title: String,
     val category: TaskCategory,
-    val accumulatedCount: Long,
-    val totalCount: Long,
+    val accumulatedCount: Int,
+    val totalCount: Int,
     val phrase: String,
     val chatStatus: Boolean, //whether somebody is in the chat room
     val taskStatus: Boolean //whether finish the task today
